# docs/swagger.yaml
basePath: /api
definitions:
  APIError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  APIErrors:
    properties:
      code:
        type: integer
      errors:
        items:
          $ref: '#/definitions/APIError'
        type: array
    type: object
...
  AdminOrderResponse:
    properties:
      cancelled_at:
        type: string
      comment:
        type: string
      created_at:
        type: string
      currency:
        type: string
      discount_total:
        type: number
      email:
        type: string
      grand_total:
        type: number
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/OrderItemResponse'
        type: array
      number:
        type: integer
      paid_at:
        type: string
      payment_method:
        type: string
      payment_status:
        type: string
      phone:
        type: string
      shipping:
        $ref: '#/definitions/OrderShippingResponse'
      shipping_total:
        type: number
      status:
        type: string
      subtotal:
        type: number
      tax_total:
        type: number
      user_id:
        type: string
    type: object
...
  CreateOrderRequest:
    properties:
      comment:
        maxLength: 2000
        type: string
      email:
        type: string
      expected_total:
        description: |-
          ExpectedTotal, when set, must equal the server-computed grand total or the
          order is rejected (price/availability changed since the cart was shown).
          Decimal string, e.g. "1990.00".
        type: string
      payment_method:
        maxLength: 32
        type: string
      phone:
        maxLength: 32
        type: string
      ship_address:
        maxLength: 512
        type: string
      ship_city_id:
        type: string
      ship_city_name:
        maxLength: 255
        type: string
      ship_postcode:
        maxLength: 16
        type: string
      ship_recipient:
        maxLength: 255
        type: string
      shipping_method:
        maxLength: 64
        type: string
    required:
    - payment_method
    - ship_address
    - ship_city_name
    - ship_recipient
    type: object
...
  FullPagingData:
    properties:
      last_page:
        type: integer
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  JSONResponse-AdminOrderResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/AdminOrderResponse'
      message:
        type: string
    type: object
...
  JSONResponse-OrderResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/OrderResponse'
      message:
        type: string
    type: object
...
  JSONResponse-ResponseWithFullPagination-AdminOrderResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/ResponseWithFullPagination-AdminOrderResponse'
      message:
        type: string
    type: object
...
  JSONResponse-ResponseWithFullPagination-OrderResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/ResponseWithFullPagination-OrderResponse'
      message:
        type: string
    type: object
...
  OrderItemResponse:
    properties:
      image_path:
        type: string
      line_total:
        type: number
      name:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
      sku:
        type: string
      slug:
        type: string
      unit_price:
        type: number
      variant_id:
        type: string
    type: object
  OrderResponse:
    properties:
      cancelled_at:
        type: string
      comment:
        type: string
      created_at:
        type: string
      currency:
        type: string
      discount_total:
        type: number
      email:
        type: string
      grand_total:
        type: number
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/OrderItemResponse'
        type: array
      number:
        type: integer
      paid_at:
        type: string
      payment_method:
        type: string
      payment_status:
        type: string
      phone:
        type: string
      shipping:
        $ref: '#/definitions/OrderShippingResponse'
      shipping_total:
        type: number
      status:
        type: string
      subtotal:
        type: number
      tax_total:
        type: number
    type: object
  OrderShippingResponse:
    properties:
      address:
        type: string
      city_id:
        type: string
      city_name:
        type: string
      method:
        type: string
      postcode:
        type: string
      recipient:
        type: string
    type: object
...
  ResponseWithFullPagination-AdminOrderResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/AdminOrderResponse'
        type: array
      pagination:
        $ref: '#/definitions/FullPagingData'
    type: object
...
  ResponseWithFullPagination-OrderResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/OrderResponse'
        type: array
      pagination:
        $ref: '#/definitions/FullPagingData'
    type: object
...
info:
  contact:
    email: go-store@stick.sh
    name: Vladislav B
  description: This is an API for go-store
  title: GO-store
  version: "1.0"
paths:
...
  /v1/admin/orders:
    get:
      consumes:
      - application/json
      description: Admin order list. Every filter is optional.
      parameters:
      - description: |-
          CreatedFrom, CreatedTo bound created_at, inclusive. RFC3339, e.g.
          "2025-01-01T00:00:00Z".
        in: query
        name: created_from
        type: string
      - in: query
        name: created_to
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      - enum:
        - unpaid
        - paid
        - refunded
        - failed
        in: query
        name: payment_status
        type: string
      - enum:
        - pending
        - paid
        - processing
        - shipped
        - delivered
        - cancelled
        - refunded
        in: query
        name: status
        type: string
      - in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-ResponseWithFullPagination-AdminOrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - Admin Order
  /v1/admin/orders/{number}:
    get:
      consumes:
      - application/json
      description: One order by its number, no ownership check
      parameters:
      - description: Order number
        in: path
        name: number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-AdminOrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrors'
      security:
      - BearerAuth: []
      summary: Get order
      tags:
      - Admin Order
...
  /v1/orders:
    get:
      consumes:
      - application/json
      description: Order history for the authenticated account
      parameters:
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-ResponseWithFullPagination-OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrors'
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - Order
    post:
      consumes:
      - application/json
      description: 'Checkout: converts the cart (session or account) into an order,
        decrements stock and clears the cart. Send an `Idempotency-Key` header (UUID)
        to make retries safe.'
      parameters:
      - description: Idempotency key (UUID)
        in: header
        name: Idempotency-Key
        type: string
      - description: Checkout data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/APIErrors'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/APIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Create order
      tags:
      - Order
  /v1/orders/{number}:
    get:
      consumes:
      - application/json
      description: One order by its number, scoped to the authenticated account
      parameters:
      - description: Order number
        in: path
        name: number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrors'
      security:
      - BearerAuth: []
      summary: Get order
      tags:
      - Order
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"